+++ HISTORY.rst
@@ -14,6 +14,8 @@ Pending Release
   ``importlib.metadata.version("pytest-randomly")``
   (`docs <https://docs.python.org/3.8/library/importlib.metadata.html#distribution-versions>`__ /
   `backport <https://pypi.org/project/importlib-metadata/>`__).
+* Convert reading entrypoints to use ``importlib.metadata``. Depend on
+  ``importlib-metadata`` on Python < 3.8.
 
 3.1.0 (2019-08-25)
 ------------------